fix(macos): cast NS arrow key constants to UInt32 in scalar set
`NSUpArrowFunctionKey` and friends are typed as `Int`, but the
`arrowKeyScalars` set is declared as `Set<UInt32>` so it can be
compared against `UnicodeScalar.value`. Swift's array literal type
inference is strict — unlike `==` comparisons, which use `BinaryInteger`
overloads — so the literal needs explicit `UInt32(...)` casts.

+++ clients/macos/vellum-assistant/App/AppDelegate+InputMonitors.swift
@@ -568,8 +568,8 @@ extension AppDelegate {
         // adds the flag. For non-arrow keys, keep .function so that explicit
         // fn+key bindings (e.g. fn+cmd+k) are not triggered without Fn held.
         let arrowKeyScalars: Set<UInt32> = [
-            NSUpArrowFunctionKey, NSDownArrowFunctionKey,
-            NSLeftArrowFunctionKey, NSRightArrowFunctionKey
+            UInt32(NSUpArrowFunctionKey), UInt32(NSDownArrowFunctionKey),
+            UInt32(NSLeftArrowFunctionKey), UInt32(NSRightArrowFunctionKey)
         ]
         func isArrowKey(_ key: String) -> Bool {
             guard let scalar = key.unicodeScalars.first, key.unicodeScalars.count == 1 else { return false }
